backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg[binary]>=3.2.10,<3.4
python-decouple==3.8
Pillow>=12.0.0,<13.0.0
drf-spectacular==0.27.2
bcrypt==4.2.1
gunicorn==23.0.0
whitenoise==6.8.2
